pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "glyph-engine"
version = "1.0.0"
description = "Symbolic control and continuity interface for human-AI interaction"
readme = "README.md"
license = "MIT"
authors = [{ name = "Paul Desai", email = "paul@n1intelligence.com" }]
requires-python = ">=3.9"
keywords = ["glyph", "ai", "governance", "mirrordna", "lingos"]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio",
]
api = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
]
all = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
]

[project.scripts]
glyph = "glyph_engine.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["glyph_engine"]
